requirements.txt
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.8
python-dotenv==1.0.0
requests==2.31.0
setuptools==65.5.0